Move sns message work to a generic folder
This is now used in multiple places
